rename shared to builder
